+++ subgraph-mainnet-production.yaml
@@ -1,5 +1,8 @@
 specVersion: 0.0.2
 description: Tinlake Mainnet Production
+graft:
+  base: QmSqT1v8x4xr2PkHortdrNo4uZtJcvq9m3N7Wo3CKEr2f3
+  block: 11062000
 schema:
   file: ./schema.graphql
 dataSources:
